Update waltz/ballroom/setup.py
Fixed start_sessions s.t. the initializer takes a default_session dict, rather than a lambda (since the patch to web.py to deal with lambdas for the initializer hasn't been accepted yet). Also, added kwargs to setup process so DiskStore can be overriden for sessions (and a db used instead of flatfile)